# [SPARK-40716][SQL] Relax the restrictions of broadcast join on bloom filter