--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-modified/q3.sf100/simplified.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-modified/q3.sf100/simplified.txt
@@ -10,6 +10,16 @@ TakeOrderedAndProject [d_year,sum_agg,brand_id,brand]
                   Project [ss_net_profit,ss_sold_date_sk,i_brand_id,i_brand]
                     BroadcastHashJoin [ss_item_sk,i_item_sk]
                       Filter [ss_item_sk]
+                        Subquery #2
+                          ObjectHashAggregate [buf] [bloom_filter_agg(xxhash64(i_item_sk, 42), 199, 1592, 0, 0),bloomFilter,buf]
+                            Exchange #3
+                              ObjectHashAggregate [i_item_sk] [buf,buf]
+                                WholeStageCodegen (1)
+                                  Project [i_item_sk]
+                                    Filter [i_manufact_id,i_item_sk]
+                                      ColumnarToRow
+                                        InputAdapter
+                                          Scan parquet spark_catalog.default.item [i_item_sk,i_manufact_id]
                         ColumnarToRow
                           InputAdapter
                             Scan parquet spark_catalog.default.store_sales [ss_item_sk,ss_net_profit,ss_sold_date_sk]
@@ -22,7 +32,7 @@ TakeOrderedAndProject [d_year,sum_agg,brand_id,brand]
                                           InputAdapter
                                             Scan parquet spark_catalog.default.date_dim [d_date_sk,d_year,d_moy]
                       InputAdapter
-                        BroadcastExchange #3
+                        BroadcastExchange #4
                           WholeStageCodegen (1)
                             Project [i_item_sk,i_brand_id,i_brand]
                               Filter [i_manufact_id,i_item_sk]