@cloud-fan 2022-10-13
what's the result of TPCDS?