@beliefer 2022-10-13
> what's the result of TPCDS?

I'm collecting.